separate immediate handling from ast structure walk
